# api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  user_auth.registerRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      phone:
        type: string
    type: object
  user_auth.sendOTPRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  user_auth.verifyOTPRequest:
    properties:
      email:
        type: string
      otp:
        type: string
    type: object
...
host: localhost:3000
info:
  contact:
    email: support@example.com
    name: MarketKit
  description: REST API for the MarketKit marketplace platform. Admin routes require
    a Bearer JWT token. User routes require a separate User Bearer JWT token.
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: MarketKit API
  version: "1.0"
paths:
...
  /auth/avatar:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Avatar file (jpg, png, webp only)
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - UserAuth: []
      summary: Upload profile picture
      tags:
      - User Auth
  /auth/change-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Current and new password
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - AdminAuth: []
      summary: Change current admin password
      tags:
      - Admin Auth
    put:
      consumes:
      - application/json
      parameters:
      - description: Current and new password
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - UserAuth: []
      summary: Change user password
      tags:
      - User Auth
...
  /auth/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - UserAuth: []
      summary: Get current user profile
      tags:
      - User Auth
    patch:
      consumes:
      - application/json
      parameters:
      - description: Profile fields
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - AdminAuth: []
      summary: Update current admin profile
      tags:
      - Admin Auth
...
  /auth/refresh:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Refresh access token
      tags:
      - User Auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: User registration payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/user_auth.registerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - User Auth
...
  /auth/send-otp:
    post:
      consumes:
      - application/json
      parameters:
      - description: Email and password payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/user_auth.sendOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Send OTP for verification
      tags:
      - User Auth
  /auth/verify-otp:
    post:
      consumes:
      - application/json
      parameters:
      - description: Email and OTP payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/user_auth.verifyOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Verify OTP and login
      tags:
      - User Auth
...
securityDefinitions:
  AdminAuth:
    description: Admin JWT — prefix with "Bearer "
    in: header
    name: Authorization
    type: apiKey
  UserAuth:
    description: User JWT — prefix with "Bearer "
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"